Also convert sinks test
