+++ packages/test/src/test-integration-workflows.ts
@@ -15,7 +15,7 @@ import { activityStartedSignal } from './workflows/definitions';
 import * as workflows from './workflows';
 import { Context, helpers, makeTestFunction } from './helpers-integration';
 import { overrideSdkInternalFlag } from './mock-internal-flags';
-import { RUN_TIME_SKIPPING_TESTS } from './helpers';
+import { RUN_TIME_SKIPPING_TESTS, asSdkLoggerSink } from './helpers';
 
 const test = makeTestFunction({ workflowsPath: __filename, workflowInterceptorModules: [__filename] });
 
@@ -987,6 +987,53 @@ test('Workflow can upsert memo', async (t) => {
       hotel: 'bar6',
     });
   });
+});
+
+test('Sink functions contains upserted memo', async (t) => {
+  const { createWorker, executeWorkflow } = helpers(t);
+  const recordedMessages = Array<{ message: string; memo: Record<string, unknown> | undefined }>();
+  const sinks = asSdkLoggerSink(async (info, message, _attrs) => {
+    recordedMessages.push({
+      message,
+      memo: info.memo,
+    });
+  });
+  const worker = await createWorker({ sinks });
+  await worker.runUntil(async () => {
+    await executeWorkflow(upsertAndReadMemo, {
+      memo: {
+        note1: 'aaa',
+        note2: 'bbb',
+        note4: 'eee',
+      },
+      args: [
+        {
+          note2: 'ccc',
+          note3: 'ddd',
+          note4: null,
+        },
+      ],
+    });
+  });
+
+  t.deepEqual(recordedMessages, [
+    {
+      message: 'Workflow started',
+      memo: {
+        note1: 'aaa',
+        note2: 'bbb',
+        note4: 'eee',
+      },
+    },
+    {
+      message: 'Workflow completed',
+      memo: {
+        note1: 'aaa',
+        note2: 'ccc',
+        note3: 'ddd',
+      },
+    },
+  ]);
 });
 
 export const interceptors: workflow.WorkflowInterceptorsFactory = () => {
